+++ package.json
@@ -13,6 +13,7 @@
     "@react-navigation/bottom-tabs": "^5.8.0",
     "@react-navigation/native": "^5.7.3",
     "@ui-kitten/components": "^5.0.0",
+    "axios": "^0.20.0",
     "expo": "^38.0.10",
     "expo-status-bar": "^1.0.2",
     "react": "~16.11.0",
@@ -22,6 +23,7 @@
     "react-native-dashboard": "0.0.4",
     "react-native-elements": "^2.3.2",
     "react-native-gesture-handler": "^1.6.1",
+    "react-native-modal": "^11.5.6",
     "react-native-reanimated": "^1.9.0",
     "react-native-safe-area-context": "^3.0.7",
     "react-native-screens": "^2.9.0",
