+++ src/components/AnonymousHeader.js
