+++ src/components/AnonymousModal.js
@@ -6,20 +6,16 @@ export default class AnonymousModal extends React.Component {
     return (
       <View
         style={{
+          position: "absolute",
+          bottom: 40,
           flex: 1,
           backgroundColor: "#fff",
           width: Dimensions.get("screen").width * 0.9,
+          flexDirection: "column",
         }}
       >
-        <View
-          style={{
-            flexDirection: "column",
-            backgroundColor: "lightblue",
-          }}
-        >
-          <Button title="匿名解除を希望する" />
-          <Button title="キャンセル" onPress={this.props.nav.toggleModal} />
-        </View>
+        <Button title="匿名解除を希望する" />
+        <Button title="キャンセル" onPress={this.props.nav.toggleModal} />
       </View>
     );
   }
